Trying to reduce possibility of inadvertently importing the SVGs
