+++ packages/framework/esm-styleguide/src/icons/index.ts
@@ -1,77 +1,1 @@
-/** @category Icons */
-import activity from './svgs/activity.svg';
-import add from './svgs/add.svg';
-import arrowDown from './svgs/arrow-down.svg';
-import arrowLeft from './svgs/arrow-left.svg';
-import arrowRight from './svgs/arrow-right.svg';
-import arrowUp from './svgs/arrow-up.svg';
-import chartAverage from './svgs/chart-average.svg';
-import chemistry from './svgs/chemistry.svg';
-import chevronDown from './svgs/chevron-down.svg';
-import chevronLeft from './svgs/chevron-left.svg';
-import chevronRight from './svgs/chevron-right.svg';
-import chevronUp from './svgs/chevron-up.svg';
-import close from './svgs/close.svg';
-import closeFilled from './svgs/close--filled.svg';
-import closeOutline from './svgs/close--outline.svg';
-import download from './svgs/download.svg';
-import edit from './svgs/edit.svg';
-import imageMedical from './svgs/image-medical.svg';
-import listChecked from './svgs/list-checked.svg';
-import location from './svgs/location.svg';
-import microscope from './svgs/microscope.svg';
-import money from './svgs/money.svg';
-import reset from './svgs/reset.svg';
-import save from './svgs/save.svg';
-import search from './svgs/search.svg';
-import shoppingCart from './svgs/shopping-cart.svg';
-import shoppingCartArrowDown from './svgs/shopping-cart--arrow-down.svg';
-import switcher from './svgs/switcher.svg';
-import time from './svgs/time.svg';
-import tools from './svgs/tools.svg';
-import translate from './svgs/translate.svg';
-import trashCan from './svgs/trash-can.svg';
-import userAvatar from './svgs/user-avatar.svg';
-import userXray from './svgs/user-xray.svg';
-import warning from './svgs/warning.svg';
-import { addSvg } from '../svg-utils';
-
 export * from './icons';
-
-export function setupIcons() {
-  addSvg('omrs-icon-activity', activity);
-  addSvg('omrs-icon-add', add);
-  addSvg('omrs-icon-arrow-down', arrowDown);
-  addSvg('omrs-icon-arrow-left', arrowLeft);
-  addSvg('omrs-icon-arrow-right', arrowRight);
-  addSvg('omrs-icon-arrow-up', arrowUp);
-  addSvg('omrs-icon-chart-average', chartAverage);
-  addSvg('omrs-icon-chemistry', chemistry);
-  addSvg('omrs-icon-chevron-down', chevronDown);
-  addSvg('omrs-icon-chevron-left', chevronLeft);
-  addSvg('omrs-icon-chevron-right', chevronRight);
-  addSvg('omrs-icon-chevron-up', chevronUp);
-  addSvg('omrs-icon-close', close);
-  addSvg('omrs-icon-close-filled', closeFilled);
-  addSvg('omrs-icon-close-outline', closeOutline);
-  addSvg('omrs-icon-download', download);
-  addSvg('omrs-icon-edit', edit);
-  addSvg('omrs-icon-image-medical', imageMedical);
-  addSvg('omrs-icon-list-checked', listChecked);
-  addSvg('omrs-icon-location', location);
-  addSvg('omrs-icon-microscope', microscope);
-  addSvg('omrs-icon-money', money);
-  addSvg('omrs-icon-reset', reset);
-  addSvg('omrs-icon-search', search);
-  addSvg('omrs-icon-save', save);
-  addSvg('omrs-icon-shopping-cart', shoppingCart);
-  addSvg('omrs-icon-shopping-cart--arrow-down', shoppingCartArrowDown);
-  addSvg('omrs-icon-switcher', switcher);
-  addSvg('omrs-icon-time', time);
-  addSvg('omrs-icon-tools', tools);
-  addSvg('omrs-icon-translate', translate);
-  addSvg('omrs-icon-trash-can', trashCan);
-  addSvg('omrs-icon-user-avatar', userAvatar);
-  addSvg('omrs-icon-user-xray', userXray);
-  addSvg('omrs-icon-warning', warning);
-}
