+++ packages/framework/esm-styleguide/src/index.ts
@@ -1,9 +1,9 @@
-import { setupLogo } from './logo';
-import { setupIcons } from './icons';
-import { setupBranding } from './brand';
 import { defineConfigSchema } from '@openmrs/esm-config';
+import { setupLogo } from './logo';
+import { setupIcons } from './icons/icon-registration';
+import { setupBranding } from './brand';
 import { esmStyleGuideSchema } from './config-schema';
-import { setupPictograms } from './pictograms';
+import { setupPictograms } from './pictograms/pictogram-registration';
 
 export * from './breakpoints';
 export * from './spinner';
@@ -19,7 +19,8 @@ export * from './responsive-wrapper';
 export * from './patient-banner';
 export * from './patient-photo';
 export * from './custom-overflow-menu';
-export * from './icons/icons';
+export * from './icons';
+export * from './pictograms';
 
 defineConfigSchema('@openmrs/esm-styleguide', esmStyleGuideSchema);
 setupBranding();
