+++ packages/framework/esm-styleguide/src/pictograms/index.ts
@@ -1,5 +1,1 @@
-import { addSvg } from '../svg-utils';
-
 export * from './pictograms';
-
-export function setupPictograms() {}
