+++ packages/framework/esm-styleguide/src/pictograms/pictogram-registration.ts
@@ -1,0 +1,3 @@
+import { addSvg } from '../svg-utils';
+
+export function setupPictograms() {}
